Fix handling of large objects in SDO Upload with Complete Access

+++ soes/esc_coe.c
@@ -468,7 +468,7 @@ static uint32_t complete_access_subindex_loop(const _objd *objd,
 
    while (nsub <= SDOobjects[nidx].maxsub)
    {
-      uint8_t bitlen = (objd + nsub)->bitlength;
+      uint16_t bitlen = (objd + nsub)->bitlength;
       void *ul_source = ((objd + nsub)->data != NULL) ?
             (objd + nsub)->data : (void *)&((objd + nsub)->value);
       uint8_t bitoffset = size % 8;
